+++ src/Library.Domain/Entities/Book.cs
@@ -5,7 +5,7 @@ public class Book
     public Guid Id { get; private set; }
     public string Title { get; private set; }
     public string Description { get; private set; }
-    public short GenreId { get; private set; }
+    public int GenreId { get; private set; }
     public virtual Genre Genre { get; private set; }
     public Guid AuthorId { get; private set; }
     public virtual Author Author { get; private set; }
@@ -13,7 +13,7 @@ public class Book
     public int Amount { get; private set; }
     public virtual ICollection<TakenBook> TakenBooks { get; private set; }
 
-    public Book(Guid? id, string title, string description, short genreId, Guid authorId, string isbn, int amount)
+    public Book(Guid? id, string title, string description, int genreId, Guid authorId, string isbn, int amount)
     {
         Id = id ?? Guid.NewGuid();
         Title = title;
